+++ app/trade/page.tsx
@@ -5,7 +5,7 @@ import { faHeart as faHeartSolid } from '@fortawesome/free-solid-svg-icons';
 import { faHeart as faHeartRegular } from '@fortawesome/free-regular-svg-icons';
 import Link from 'next/link';
 import React, { useEffect, useState } from 'react';
-import { TradeData } from '@/app/_components/communityComponents/tradeSupabase';
+import { TradeData } from '@/app/_api/tradeApi';
 import { formatToLocaleDateTimeString } from '../_utils/date';
 import Image from 'next/image';
